Rename studyjobcontroller to studyjob
