+++ pkg/controller/studyjob/model_util.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package studyjobcontroller
+package studyjob
 
 import (
 	"context"
